Consolidate dimension handling helpers into a shared utility module
Moves common dimension defaulting and validation logic, previously duplicated across various processing modules, into the new `spinlab.processing._utils` module.

This refactoring removes individual `_get_..._dim` helpers from modules like `align`, `apodization`, `average`, `integration`, `interpolation`, `offset`, and `phase`, replacing them with a single import from `_utils.get_default_dim`.

`spinlab/processing/helpers.py` is repurposed to serve primarily as a backward-compatibility export layer, now also exposing the new utility functions.
